chore: rename passwordbox to password-box [x] no test file to update
