+++ packages/cfd/src/Containers/dmt5-trade-modal.tsx
@@ -14,7 +14,7 @@ import {
 import { Localize, localize } from '@deriv/translations';
 import { getPlatformMt5DownloadLink } from '../Helpers/constants';
 import SpecBox from '../Components/specbox';
-import PasswordBox from '../Components/passwordbox';
+import PasswordBox from '../Components/password-box';
 import TradingPlatformIcon from '../Assets/svgs/trading-platform';
 import { TTradingPlatformAccounts } from '../Components/props.types';
